fix(f20.a.1): L2CAP bounds guard — avoid uint8_t overflow on large l2_len
Code review of a43a53c. The guard length >= (uint8_t)(l2_len + 4u)
silently disabled itself for l2_len > 251 because the cast wraps.
Use (uint16_t)length >= (l2_len + 4u) so the comparison is honest
regardless of payload size.

+++ firmware/cc1352/src/ble20_dispatch.c
@@ -35,7 +35,7 @@ void Ble20_drainAndDispatch(uint8_t *reason_out) {
             /* L2CAP frame: [len:2 LE][cid:2 LE][payload] */
             uint16_t l2_len = (uint16_t)pkt[2] | ((uint16_t)pkt[3] << 8);
             uint16_t l2_cid = (uint16_t)pkt[4] | ((uint16_t)pkt[5] << 8);
-            if (l2_cid == 0x0004u && l2_len >= 1u && length >= (uint8_t)(l2_len + 4u)) {
+            if (l2_cid == 0x0004u && l2_len >= 1u && (uint16_t)length >= (l2_len + 4u)) {
                 AttServer_handleRequest(&pkt[6], (uint8_t)l2_len);
             }
         }
